ML_Service/requirements.txt
# =============================================================================
# AIVerifySnap ML Service - Production Dependencies
# =============================================================================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Image Processing
pillow>=11.0.0
numpy>=2.0.0
opencv-python-headless>=4.9.0

# PyTorch (CPU version - use torch-cuda for GPU support)
# For CUDA support, install: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.2.0
torchvision>=0.17.0

# Data validation
pydantic>=2.0.0

# HTTP client (for potential external service calls)
httpx>=0.27.0

# Logging and monitoring
python-json-logger>=2.0.0
